Moving the UnitOfWork to Event Store
