+++ src/Services/Account/Command/Application/Abstractions/IEventStoreGateway.cs
@@ -1,10 +1,11 @@
 using Domain.Abstractions.Aggregates;
 
-namespace Application.Abstractions.Gateways;
+namespace Application.Abstractions;
 
 public interface IEventStoreGateway
 {
     Task AppendEventsAsync(IAggregateRoot aggregate, CancellationToken cancellationToken);
     Task<TAggregate> LoadAggregateAsync<TAggregate>(Guid aggregateId, CancellationToken cancellationToken) where TAggregate : IAggregateRoot, new();
     IAsyncEnumerable<Guid> StreamAggregatesId(CancellationToken cancellationToken);
+    Task ExecuteTransactionAsync(Func<CancellationToken, Task> operationAsync, CancellationToken cancellationToken);
 }
