--- src/Services/Account/Command/Application/Abstractions/IUnitOfWork.cs
@@ -1,6 +1,0 @@
-namespace Application.Abstractions;
-
-public interface IUnitOfWork
-{
-    Task ExecuteAsync(Func<CancellationToken, Task> operationAsync, CancellationToken cancellationToken);
-}
