+++ src/Services/Account/Command/Application/Services/ApplicationService.cs
@@ -1,5 +1,4 @@
 ﻿using Application.Abstractions;
-using Application.Abstractions.Gateways;
 using Contracts.Abstractions.Messages;
 using Domain.Abstractions.Aggregates;
 
@@ -9,16 +8,13 @@ public class ApplicationService : IApplicationService
 {
     private readonly IEventStoreGateway _eventStoreGateway;
     private readonly IEventBusGateway _eventBusGateway;
-    private readonly IUnitOfWork _unitOfWork;
 
     public ApplicationService(
         IEventStoreGateway eventStoreGateway,
-        IEventBusGateway eventBusGateway,
-        IUnitOfWork unitOfWork)
+        IEventBusGateway eventBusGateway)
     {
         _eventStoreGateway = eventStoreGateway;
         _eventBusGateway = eventBusGateway;
-        _unitOfWork = unitOfWork;
     }
 
     public Task<TAggregate> LoadAggregateAsync<TAggregate>(Guid id, CancellationToken cancellationToken)
@@ -26,7 +22,7 @@ public class ApplicationService : IApplicationService
         => _eventStoreGateway.LoadAggregateAsync<TAggregate>(id, cancellationToken);
 
     public Task AppendEventsAsync(IAggregateRoot aggregate, CancellationToken cancellationToken)
-        => _unitOfWork.ExecuteAsync(
+        => _eventStoreGateway.ExecuteTransactionAsync(
             operationAsync: async ct =>
             {
                 await _eventStoreGateway.AppendEventsAsync(aggregate, ct);
