+++ src/Services/Account/Command/Infrastructure.EventStore/DependencyInjection/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
 using Application.Abstractions;
-using Application.Abstractions.Gateways;
 using Domain.Abstractions.EventStore;
 using Infrastructure.EventStore.Contexts;
 using Infrastructure.EventStore.DependencyInjection.Options;
@@ -16,7 +15,6 @@ public static class ServiceCollectionExtensions
     {
         services.AddScoped<IEventStoreGateway, EventStoreGateway>();
         services.AddScoped<IEventStoreRepository, EventStoreRepository>();
-        services.AddScoped<IUnitOfWork, UnitOfWork>();
 
         services.AddDbContextPool<EventStoreDbContext>((provider, builder) =>
         {
